Move NewRawFromZeekTSV from zbuf to zeekio
